Review and polish the new support for svn and darcs, and use it in some recipes.
We define el-get-svn and el-get-darcs for platforms where occasional users
won't want to adapt their PATH and still need to use odd placed binaries,
like for example "~/.cabal/bin/darcs". Using an older darcs version (1.0.9)
or newer (2.x) repositories will result in HTTP 404 error, but I guess
that's not for `el-get' to try and solve that.