Add :destination option for both http and ftp supports.
This option is useful if the package sould be retrived using a presentation
insterface (such as as web SCM tool). The file name could be specified in
the :destination part of the package configuration.

For example, destination should be set to \"package.el\" if the package url
has the following scheme:

"http://www.example.com/show-as-text?file=path/package.el"

Signed-off-by: Sébastien Gross <seb•ɑƬ•chezwam•ɖɵʈ•org>