Implement full support for :repo property on elpa sources
This allows you to specify the repository from which an elpa package
should be installed, which allows the install to work regardless of the currently-set repositories.

Obviously, this is a security hole if you use an unknown recipe
without auditing it. It may be prudent to implement a prompt when
adding any unknown package repository.