`el-get' now defaults to caring about packages known "required" and "installed" from the status file.
This should bring a much easier to understand semantics, near of that of
apt-get and friends: by default, any installed package is automatically
initialized at next startup without special care.  If you don't need local
recipes, you don't need to edit and maintain any `el-get-sources'.

It's still possible to manually prepare a list of packages to install, so
that you can share your setup between multiple installations.  To do that,
give the package list (names or symbols) to the (el-get) call in your setup.
This way you can also maintain specific lists depending on system or network
or whatever is useful for you.  That's the advanced setup, documented.