Fix a bug in el-get-init-and-install.
We were failing to find the package in the list of packages to init and so
considered them for installation again.  As el-get-install will respect the
current package status, that would result in the package not being
initialized at startup.

Bug found while reading the code to prepare the 3.1 release.