color themes depend on color-theme and are automatically required
Autoloads are rarely (if ever?) present for color themes. color-theme-zenburn
installs its own in the :post-init hook, but that's not typically done for el-get
packages, so I've just added a :load or :feature property (depending on whether
the library provides a feature).

For color-theme-twilight, the current version available automatically installs
itself as the current color theme, so the :feature property is omitted. I have a
pull request in to the project, so if that's fixed, I'll add it.