Rewrite el-get-funcall to expect only function or single lisp form
As discussed in
https://github.com/dimitri/el-get/pull/582#issuecomment-4081824

Since the lisp-form is now recommended and functions are deprecated,
rename "el-get-funcall" to el-get-run-package-support, and change
other terms in documentation and comments from "function" to form.
When it gets a function (anything callable), it gives a deprecation
warning that, when practical, tells the user how to modify the recipe
to use a lisp form instead of a function.

Also, now ":post-init 'FORM" is equivalent to ":post-init FORM",
because any top-level quoting is stripped from FORM (with a warning)
before evaluating it.