Clean elc file if it is the *same* age as el file
Before, it would only be cleaned if it was older. Now being exactly
the same age also means it gets cleaned. This is because if they are
the same age, we don't know which came first, so we just have to
assume that the elc file is stale and recreate it.