Only save and load status recipes for installed packages
If a package's status is anything other than "installed", the recipe
property is always saved as nil in the status file, and the recipe
property is not loaded from the status file at all.

This fixes a bug in migration from the old status format. If a package
was listed as removed and its recipe file was no longer available, the
status migration would fail when el-get tried to read the recipe for
that package so that it could store that recipe in the status file,
even though there was no need to do so.