Fix cedet recipe
This includes a fix to "el-get-reload" that ignores errors during
reloading package files, because CEDET raises errors if you try to
reload it. If a file raises an error during reloading, we issue a
warning telling the user that they might have to restart emacs to
finish the update.

Recipe fixes based on https://github.com/dimitri/el-get/pull/748