Remove el-get 'wait that's been broken for a long time already, fixes #758.
The use case for el-get 'wait is very thin (setup file only), and fixing it
would be hard enough that we're not going to work on it. If you want to wait
until the packages are installed, just use 'sync.

One reason why fixing that situation is not easy is that some error cases
that happen at install time are not catched by El-Get code, so there's no
place where to set the package status as "failed" or something equivalent.
We wouls still have some "required" packages left, and if we don't count
them in the progress information we hang up completely.

This commit also refresh the test/README content to cope with the recent
status.old backup file, and in passing reformat the file as Markdown.