Replaced 'org-install with 'ob-tangle when checking to see if org-babel-load-file is present.
'org-install is now deprecated in Org (current git HEAD) and will be
deprecated in the version shipped with Emacs 24.3.  The loading
function is found in `ob-tangle` however the message will prompt users
to `(require 'org)` since this will load `ob-tangle` as well.

Also fixed missing quote on package loading functions.