don't be "installed" during autoload removal
`el-get-remove' only sets package status to "removed" after performing
the removal; this uses `el-get-update-autoloads' to save the autoloads
file, which re-adds the just removed autoloads if the package is still
"installed" state.

This commit fixes #446 by setting the package state to "required" state
before performing the removal.  What's really wanted is the *opposite*
of "required" ("unrequired" sounds awkward), but that would be a much
more complicated change.