fix el-get-remove-autoloads for emacs != 24.3
Instead of tricking autoload-find-destination to do what we want, pull
out the relevant bits of code that do removal.

Using autoload-find-destination works in emacs 24.3 apparently because
of a bug: file-attributes returns a 4 integer list for time, but the
autoload function expects the old 2 integer format.  In 24.2 where
file-attributes still returns 2 integers, the autoload section is
considered up to date and is not removed.