Do not attempt to load CEDET if eieio has been loaded
CEDET will generate an error if it is loaded after the eieio version that is
shipped with Emacs has been loaded.  During bootstrapping this will continuously
cause my Emacs to stop installing packages and forcing me to restart it.

This change will ensure that Emacs can be fully bootstrapped using el-get even
if the updated version of CEDET is installed.  The new version will simply be
loaded on the next launch.