magit.rcp: use el-get's default compilation
We still use make for producing the info file. Since Windows users will
probably not go through the trouble of installing make and makeinfo,
disable that step in Windows.

The magit-autoloads file is produced in the Makefile from ;;;###autoload
cookies; since el-get extracts those and puts them into .loaddefs by
default we don't need to (require 'magit-autoloads) anymore.

Removed the ("touch" "`find . -name Makefile`") from
:build/berkeley-unix since that would create a file literally called
"`find . -name Makefile`", the `` would not be expanded as the shell is
not being invoked. It was probably some leftover that's no longer needed.