Create empty dir if no apt-get dir to symlink to
If the apt-get package doesn't contain any elisp files symlinking to the
non-existant /usr/share/emacs/site-lisp/<package> directory will leave
us with a broken link that will raise an error when we attempt to look
for autoloads, clean stale elc files and the like. In this case, simply
create an empty directory instead of a link.

* methods/el-get-apt-get.el (el-get-dpkg-symlink): call `make-directory'
  instead of symlink when `debdir' doesn't exist.