Fix build rules in mmm-mode recipe
This is a followup on the pull request that I filed recently [1].  The
hack proposed in that commit was needed for avoinding a locally
modified version.texi file interfering with the build process.  This
has been reported [2] and fixed upstream [3].  This commit takes the
upstream changes into account and replaces the cp/mv commands by "make
maintainer-clean" before the autotools/make commands.

[1] https://github.com/dimitri/el-get/pull/2042
[2] https://github.com/purcell/mmm-mode/issues/41
[3] https://github.com/purcell/mmm-mode/commit/53524db13dfedcc104866c3cd3c62adda1232679