Allow package-status-alist arg
In el-get-do-init and el-get-reload but mark it obsolete. This avoids
an error while updating el-get, since the code doing the updating still
uses the old calling convention.

For the el-get-with-status-sources macro, just keep the argument but
always pass nil when calling it (because the obsolete arg is followed
by &rest args, the calling convention can't be changed in a
distinguishable way).