Fix URLs for chuck-mode recipe
The bitbucket.org URLs are broken. as today.  The ones at GitHub used in
this commit are working for me.  The author of the chuck-mode.el file at
GitHub is Kao Cardoso Félix, who was the owner of the now defunct
Bitbucket repository, so the changes in this commit should be okay.