update helm-bibtex and add ivy-bibtex
There is now an additional package ivy-bibtex in the
tmalbsburg/helm-bibtex repository (and on MELPA). Helm-bibtex.el and ivy-bibtex.el are now two alternative
backends for the backend-independent bibtex-completion.el. The former
backend depends on helm whereas the latter depends on ivy (swiper).

So I updated the helm-bibtex and added the ivy-bibtex recipes, using the current
descriptions and dependencies.

Note that ideally, one should exclude the helm-bibtex.el file from the
ivy-bibtex package and the ivy-bibtex.el file from the helm-bibtex
package (as is done on MELPA). But I didn't find how to do this with el-get.