Prevent jump to debugger from failed assertion
In Emacs 25.1, a failed cl-assert will always jump to the debugger if
debug-on-error is non-nil.  In batch-mode runs this becomes confusing
with the message suppression we have, and the crazy long stack traces
that are printed are really hard to read, so suppress this debugger
behaviour with advice as well.