Consider :emacswiki over HTTPS as secure
Secure connection is not the same as trusted (or authenticated) content
of the package. Scope of the check for secure URL is to decide could
third-party tamper with the package's content while on transit.

If the package's content is trusted or not is a separate issue. El-get
package can be considered as trusted only with :checksum recipe
property.