Skip assert (stringp url) if el-get-allow-insecure
Some methods, :elpa for instance, need some some special handling to
figure out the URL.  In the mean time, don't throw assertion violations
if the user didn't ask to check for "insecure" packages anyway.

* el-get-methods.el (el-get-insecure-check): Move check of
`el-get-allow-insecure' to beginning, before the assertion.  Change the
parameter names to lower case, as is customary in elisp.