removes auto-complete from dependencies
bringing in auto-complete and company mode as dependencies gets me an
ensime setup that enables both for completion, and they fight for
popup rights over certain completion events, giving me garbage
onscreen ( GNU Emacs 25.1.50.2 (x86_64-pc-linux-gnu, GTK+ Version
3.10.8) )

ensime's own docs suggest that auto-complete is an optional add-on
that should be installed independently, and company mode is the
preferred and indeed default option. It makes more sense to me to
leave auto-complete out of the recipe, as this doesn't break anything,
and leave it to be an optional additional install and config, which
seems to be what the ensime installation expects.