Fix w32 elpa copy-directory call for Emacs 26+
In Emacs 26, copy-directory copies the directory itself, based on
whether NEWNAME is directory name (ends with a "/").  Earlier Emacs
versions actually checked whether NEWNAME was existing directory, so
we happened to get the right behaviour most of the time.

Pass t for COPY-CONTENTS to ensure the correct behaviour all of the
time.