Avoid errors when calling dpkg-query
When a package is not installed in the system, the dpkg-query command exits
with an error.  This will then make the function
el-get-dpkg-package-installed-p fails, which interrupts the installation of
the package, making the apt-get method useless.  This commits avoids the
problem by wrapping the call to process-lines into a condition-case
statement.