Fix for mu4e to get mu4e.info from proper location
The current recipe ends with an error "/usr/sbin/install-info el-get could not build mu4e" and if you do the trace it says that the file mu4e.info could not be found. There has been some changes apparently to the mu4e build system and so the info is actually produced with makeinfo into the build directory. The small change reflects this and everything seems to install correctly now.