Add and update recipes for org-gcal, org-fc, org-pomodoro-third-time, calfw-blocks and related dependencies (#2893)
* Update typescript-mode recipe organization

The mode has been moved to the `emacs-typescript` organization

* Add a recipe for tsi.el

`orzechowskid/tsi.el` is a minor mode to indent code using the syntax
tree generated by `elisp-tree-sitter`

* Add a recipe for apheleia

`radian-software/apheleia` is a tool to auto-format files in different
languages intelligently.

* Add a recipe for humanoid-themes

`humanoid-colors/emacs-humanoid-themes` provides a beautiful light
theme and a beautiful dark theme for Emacs GUI and terminal.

Use the themes with `(load-theme 'humanoid-light t)` and
`(load-theme 'humanoid-dark t)`

* Add a recipe for ctrlf

`radian-software/ctrlf` is a modern replacement of `isearch`, keeping
to mostly the same interface.

* Update the recipe for zig-mode

- `zig-mode` is now served from a different github organization
- Add a `minimum-emacs-version` clause

* Add a recipe for calfw-blocks

This package enables rendering of tasks as calendar time-blocks in the
Emacs Calendar Framework

* Update org-gcal recipe, add recipes for emacs-aio and emacs-oauth2-auto

The latest version adds a new dependencies for dealing with Google's
OAuth2 changes.

* Add a recipe for org-pomodoro-third-time

This package modifies the existing `org-pomodoro` to implement the
third time system. The Third Time system is explained in the README of
the project.

* Add a recipe for org-fc

`org-fc` implements a spaced repetition system similar to Anki inside
org-mode. It's quite brilliant!