fix: car for t in el-get-read-status-file-force
When I remove el-get dir and try to reinitialize,
`el-get-read-status-file-force` fails with an error
`Wrong type argument: listp, t`. This is due to `make-directory` returns
`t` when directry is aleady exists. Indeed, it actually guarantees that
retruning non-nil [1]. So, `el-get-read-status-file-force` tries to
evaluate the following sexp:

```
(let ((ps t))
  (cond
   ((null ps)
    (list (list 'el-get 'status "required")))
   ((consp (car ps)) ps)
   (t (el-get-convert-from-old-status-format ps))))
```

This patch ensures that this case is handled by the default branch.

[1] https://www.gnu.org/software/emacs/manual/html_node/elisp/Create_002fDelete-Dirs.html#index-make_002ddirectory